csc591-iot-project-web-app-master/db_watcher/requirements.txt
influxdb==5.0.0
paho-mqtt==1.3.1
